+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added new formtool apis `namespaceForm`, `namespaceList`, `namespaceOverview`, `parseForm`.
 - Added new middleware apis `setLinzNamespace`, `setModelForm`, `setRecord`.
 - Moved all public middleware to the `linz.api.middleware` namespace.
+- Fixed the datepicker in documentarray modals.
 
 ## v1.0.0-16.1.0 (9 August 2018)